TST: move portable reload test to top level
