--- databroker/_mds/portable_mds/tests/conftest.py
@@ -1,34 +1,0 @@
-import os
-import tempfile
-import shutil
-import tzlocal
-import pytest
-import databroker.headersource.mongoquery as mqmds
-
-from ....headersource import sqlite as sqlmds
-
-variations = [mqmds,
-              sqlmds]
-
-
-@pytest.fixture(params=variations, scope='function')
-def mds_all(request):
-    '''Provide a function level scoped FileStore instance talking to
-    temporary database on localhost:27017 with both v0 and v1.
-
-    '''
-    tempdirname = tempfile.mkdtemp()
-    mds = request.param.MDS({'directory': tempdirname,
-                             'timezone': tzlocal.get_localzone().zone}, version=1)
-    filenames = ['run_starts.json', 'run_stops.json', 'event_descriptors.json',
-                 'events.json']
-    for fn in filenames:
-        with open(os.path.join(tempdirname, fn), 'w') as f:
-            f.write('[]')
-
-    def delete_dm():
-        shutil.rmtree(tempdirname)
-
-    request.addfinalizer(delete_dm)
-
-    return mds
